pointerevents/pointerevent_capture.html

<!doctype html>
<html>
    <head>
        <title>Set/Release capture</title>
        <meta name="viewport" content="width=device-width">
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
    </head>
    <body>
        <h1>Pointer Events capture test</h1>
        <h4>
            Test Description: This test checks if setCapture/releaseCapture functions works properly. Complete the following actions:
            <ol>
                <li> Put your mouse over the upper rectangle. pointerover event should be received for the black rectangle</li>
                <li> Put your mouse over the lower rectangle. pointerover event should be received for the purple rectangle</li>
                <li> Press and hold left mouse button over "Set Capture" button</li>
                <li> Put your mouse over the upper rectangle. pointerover should be received for the black rectangle and "gotpointercapture" should be logged inside of it</li>
                <li> Put your mouse over the lower rectangle. pointerover should be received for the black rectangle</li>
                <li> Release left mouse button. "lostpointercapture" should be logged inside of the black rectangle</li>
            </ol>
        </h4>
        Test passes if the proper behaviour of the events is observed.
        <div id="target0"></div>
        <br>
        <div id="target1"></div>
        <br>
        <input type="button" id="btnCapture" value="Set Capture">
        <script type='text/javascript'>
            var isPointerCapture = false;
            var count=0;
            
            var target0 = document.getElementById('target0');
            var target1 = document.getElementById('target1');
            var captureButton = document.getElementById('btnCapture');
            
            var test_gotpointercapture = async_test("gotpointercapture event received");
            var test_lostpointercapture = async_test("lostpointercapture event received");
            
            window.onload = function() {
            	captureButton.addEventListener('pointerdown', function(e) {
            			if(isPointerCapture == false) {
            				isPointerCapture = true;
            				captureButton.value = 'Release Capture';
            				sPointerCapture(e);
            			}
            			else {
            				rPointerCapture(e);
            			}
            		});
            		
            	target0.addEventListener('gotpointercapture', function(e) {
            		test_gotpointercapture.done();
            		log("gotpointercapture", document.getElementById('target0'));
            	});
            		
            	target0.addEventListener('lostpointercapture', function(e) {
            		test_lostpointercapture.done();
            		log("lostpointercapture", document.getElementById('target0'));
            		captureButton.value = 'Set Capture';
            		isPointerCapture = false;
            	});
            		
            	run();
            }
            
            function run() {
            	var test_pointerover0 = async_test("pointerover event for black rectangle received")
            	var test_pointerover1 = async_test("pointerover event for purple rectangle received")
            	
            	on_event(target0, "pointerover", function (event) {
            		test_pointerover0.done();	
            		log("pointerover", document.getElementById('target0'));
            		if(isPointerCapture) {
            			test(function() {
            				assert_true(event.relatedTarget==null, "relatedTarget is null when the capture is set")
            			}, "relatedTarget is null when the capture is set. relatedTarget is " + event.relatedTarget);
            		}
            	});
            	
            	on_event(target1, "pointerover", function (event) {
            		test(function() {
            			assert_true(isPointerCapture==false, "pointerover shouldn't trigger for this target when capture is enabled");
            		}, "pointerover shouldn't trigger for the purple rectangle while the black rectangle has capture");				
            		
            		test_pointerover1.done();
            		log("pointerover", document.getElementById('target1'));
            	});	
            }
        </script>
        <h1>Pointer Events Capture Test</h1>
        <div id="complete-notice"></div>
        <div id="log"></div>
    </body>
</html>
